Updates Create Attribute example.
