+++ examples/create_attribute/my_vehicle.py
@@ -1,0 +1,84 @@
+"""
+my_vehicle.py:
+
+Custom Vehicle subclass to add IMU data.
+"""
+
+from dronekit import connect, Vehicle, VehicleMode
+from pymavlink import mavutil
+import time
+
+class RawIMU(object):
+    """
+    The RAW IMU readings for the usual 9DOF sensor setup. 
+    This contains the true raw values without any scaling to allow data capture and system debugging.
+    
+    The message definition is here: https://pixhawk.ethz.ch/mavlink/#RAW_IMU
+    
+    :param time_boot_us: Timestamp (microseconds since system boot). #Note, not milliseconds as per spec
+    :param xacc: X acceleration (mg)
+    :param yacc: Y acceleration (mg)
+    :param zacc: Z acceleration (mg)
+    :param xgyro: Angular speed around X axis (millirad /sec)
+    :param ygyro: Angular speed around Y axis (millirad /sec)
+    :param zgyro: Angular speed around Z axis (millirad /sec)
+    :param xmag: X Magnetic field (milli tesla)
+    :param ymag: Y Magnetic field (milli tesla)
+    :param zmag: Z Magnetic field (milli tesla)    
+    """
+    def __init__(self, time_boot_us=None, xacc=None, yacc=None, zacc=None, xygro=None, ygyro=None, zgyro=None, xmag=None, ymag=None, zmag=None):
+        """
+        RawIMU object constructor.
+        """
+        self.time_boot_us = time_boot_us
+        self.xacc = xacc
+        self.yacc = yacc
+        self.zacc = zacc
+        self.xgyro = zgyro
+        self.ygyro = ygyro
+        self.zgyro = zgyro
+        self.xmag = xmag        
+        self.ymag = ymag
+        self.zmag = zmag      
+        
+    def __str__(self):
+        """
+        String representation used to print the RawIMU object. 
+        """
+        return "RAW_IMU: time_boot_us={},xacc={},yacc={},zacc={},xgyro={},ygyro={},zgyro={},xmag={},ymag={},zmag={}".format(self.time_boot_us, self.xacc, self.yacc,self.zacc,self.xgyro,self.ygyro,self.zgyro,self.xmag,self.ymag,self.zmag)
+
+   
+class MyVehicle(Vehicle):
+    def __init__(self, *args):
+        super(MyVehicle, self).__init__(*args)
+
+        # Create an Vehicle.raw_imu object with initial values set to None.
+        self._raw_imu = RawIMU()
+
+        # Create a message listener using the decorator.   
+        @self.on_message('RAW_IMU')
+        def listener(self, name, message):
+            """
+            The listener is called for messages that contain the string specified in the decorator,
+            passing the vehicle, message name, and the message.
+            
+            The listener writes the message to the (newly attached) ``vehicle.raw_imu`` object 
+            and notifies observers.
+            """
+            self._raw_imu.time_boot_us=message.time_usec
+            self._raw_imu.xacc=message.xacc
+            self._raw_imu.yacc=message.yacc
+            self._raw_imu.zacc=message.zacc
+            self._raw_imu.xgyro=message.xgyro
+            self._raw_imu.ygyro=message.ygyro
+            self._raw_imu.zgyro=message.zgyro
+            self._raw_imu.xmag=message.xmag
+            self._raw_imu.ymag=message.ymag
+            self._raw_imu.zmag=message.zmag
+            
+            # Notify all observers of new message (with new value)
+            self.notify_attribute_listeners('raw_imu', self._raw_imu) 
+
+    @property
+    def raw_imu(self):
+        return self._raw_imu
